refactor: move button to widgets folder
