+++ system/ui/setup.py
@@ -10,11 +10,11 @@ import pyray as rl
 from cereal import log
 from openpilot.system.hardware import HARDWARE
 from openpilot.system.ui.lib.application import gui_app, FontWeight
-from openpilot.system.ui.lib.button import gui_button, ButtonStyle
 from openpilot.system.ui.lib.label import gui_label, gui_text_box
 from openpilot.system.ui.lib.widget import Widget
-from openpilot.system.ui.widgets.network import WifiManagerUI, WifiManagerWrapper
+from openpilot.system.ui.widgets.button import gui_button, ButtonStyle
 from openpilot.system.ui.widgets.keyboard import Keyboard
+from openpilot.system.ui.widgets.network import WifiManagerUI, WifiManagerWrapper
 
 NetworkType = log.DeviceState.NetworkType
